+++ src/lob_asset_management/aux/file.clj
@@ -12,6 +12,14 @@
 
 (defn edn->file [data file-path]
   ;TODO Avoid the necessity of an existent folder
+  (io/make-parents file-path)
   (with-open [out (io/writer file-path)]
     (binding [*out* out]
       (clojure.pprint/pprint data))))
+
+(defn edn->file-table [data file-path]
+  ;TODO Avoid the necessity of an existent folder
+  (io/make-parents file-path)
+  (with-open [out (io/writer file-path)]
+    (binding [*out* out]
+      (clojure.pprint/print-table data))))